s3Events/pyproject.toml
[tool.poetry]
name = "s3events"
version = "0.1.0"
description = ""
authors = ["kevinwochan <kevinwochan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.100.0"
constructs = ">=10.0.0,<11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
